Continued on Matrix class

+++ MatrixSolver/utils.h
@@ -6,7 +6,7 @@ namespace utils {
 	const int ROWS = 3;
 	const int COLS = 4;
 
-	const int CONST_MATRIX[ROWS][COLS] = {
+	const double CONST_MATRIX[ROWS][COLS] = {
 	{1, 1, 1, 2},
 	{2, 3, 1, 3},
 	{1, -1, -2, -6}
